requirements.txt
annotated-types==0.7.0
Authlib==1.5.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dparse==0.6.4
ecdsa
filelock==3.16.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lagrange==3.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
narwhals==1.34.1
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
plotly==6.0.1
psutil==6.1.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
safety==3.3.1
safety-schemas==0.0.11
setuptools==78.1.0
shamirs==3.0.3
shellingham==1.5.4
six==1.17.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
